requirements.txt
aiohttp-jinja2==1.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
certifi==2020.11.8
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna-ssl==1.1.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
json5==0.9.5
jupyter==1.0.0
jupyter-client
jupyter-console==6.2.0
jupyterlab==2.2.6
Keras==2.3.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
matplotlib==3.0.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
notebook
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
parso==0.7.1
Pillow==5.4.1
prometheus-client==0.8.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyOpenSSL
pyparsing==2.4.7
pyreadline==2.1
python-dateutil==2.8.1
pytz==2020.4
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.20.2
scipy==1.4.1
seaborn
Send2Trash==1.5.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1